+++ lib/server/plugins/api/index.js
@@ -1,123 +1,12 @@
-var path = require('path');
-
 var _ = require('lodash');
-var Wreck = require('wreck');
 
-var hoodiejs = require('../../../helpers/pack_hoodie');
+var internals = require('./internals');
 var plugins = require('../../../helpers/plugin_api');
 
-var internals = {
-  uiKitPath: path.dirname(require.resolve('hoodie-admin-dashboard-uikit/dist/index.html')),
-  mapProxyPath: function (request, callback) {
-    //use the bearer token as the cookie AuthSession for couchdb:
-    if (request.headers.authorization && request.headers.authorization.substring(0, 'Bearer '.length) === 'Bearer ') {
-      request.headers.cookie = 'AuthSession=' + request.headers.authorization.substring('Bearer '.length);
-    } else {
-      delete request.headers.cookie;
-    }
-    request.headers.host = [internals.couchCfg.host, internals.couchCfg.port].join(':');
-    callback(null, internals.couchCfg.url + request.url.path.substr('/_api'.length), request.headers);
-  },
-  notFound: function (request, reply) {
-    reply({
-      'error': 'not found'
-    }).code(404);
-  },
-  extractToken: function (cookieHeader) {
-    var result = (/AuthSession=(.*); Version(.*)/).exec(cookieHeader[0]);
-    if (Array.isArray(result)) {
-      return result[1];
-    }
-  },
-  addCorsAndBearerToken: function (err, res, request, reply) {
-
-    if (err) {
-      reply(err).code(500);
-      return;
-    }
-
-    Wreck.read(res, {
-      json: true
-    }, function (err, data) {
-      var resp;
-      var allowedHeaders = [
-        'authorization',
-        'content-length',
-        'content-type',
-        'if-match',
-        'if-none-match',
-        'origin',
-        'x-requested-with'
-      ];
-
-      function addAllowedHeaders(arr) {
-        for (var i = 0; i < arr.length; i++) {
-          if (allowedHeaders.indexOf(arr[i].trim().toLowerCase()) === -1) {
-            allowedHeaders.push(arr[i].trim().toLowerCase());
-          }
-        }
-      }
-
-      if (err) {
-        reply(err).code(500);
-        return;
-      }
-
-      if (Array.isArray(res.headers['set-cookie'])) {
-        data.bearerToken = internals.extractToken(res.headers['set-cookie']);
-        delete res.headers['set-cookie'];
-      }
-
-      addAllowedHeaders(Object.keys(request.headers));
-
-      if (request.method === 'options') {
-        res.statusCode = 200;
-        if (request.headers['Allow-Control-Request-Headers']) {
-          addAllowedHeaders(request.headers['Allow-Control-Request-Headers'].split(','));
-        }
-      }
-
-      // hapi eats newlines. We like newlines. For POSIX and such.
-      // data = data + '\n';
-      resp = reply(data).code(res.statusCode).hold();
-      resp.headers = res.headers;
-      resp.headers['content-length'] = data ? data.length : 0;
-      resp.headers['access-control-allow-origin'] = request.headers.origin || '*';
-      resp.headers['access-control-allow-headers'] = allowedHeaders.join(', ');
-      resp.headers['access-control-expose-headers'] = 'content-type, content-length, etag';
-      resp.headers['access-control-allow-methods'] = 'GET, PUT, POST, DELETE';
-      resp.headers['access-control-allow-credentials'] = 'true';
-      resp.send();
-    });
-  }
-};
-
 exports.register = function (plugin, options, next) {
-  // FIXME: have to define these three internals here,
-  //     because they rely on `options` being defined.
-  //     This means we can't write unit tests for them,
-  //     and it's also in itself ugly. Is there a nicer
-  //     way of doing this?
-  internals.couchCfg = options.app.couch;
-  internals.handlePluginRequest = function (request, reply) {
-
-    var hooks = options.app.hooks;
-    var pluginName = request.params.name;
-    if (!pluginName) {
-      internals.notFound(request, reply);
-      return;
-    }
-
-    // the plugin is now responsible to call `reply()`
-    hooks.runDynamicForPlugin(pluginName, 'server.api.plugin-request', [request, reply]);
-  };
-  internals.getHoodiePath = function (request, reply) {
-    reply(hoodiejs(options.app))
-    .type('application/javascript')
-    .header(
-      'Pragma', 'no-cache'
-    );
-  };
+  var mapProxyPath = internals.mapProxyPath.bind(null, options.app.couch);
+  var getHoodiePath = internals.getHoodiePath.bind(null, options);
+  var handlePluginRequest = internals.handlePluginRequest.bind(null, options);
 
   // allow clients to request a gzip response, even if the
   // Accept-Encoding headers is missing or mangled due to
@@ -137,7 +26,7 @@ exports.register = function (plugin, options, next) {
       handler: {
         proxy: {
           passThrough: true,
-          mapUri: internals.mapProxyPath,
+          mapUri: mapProxyPath,
           onResponse: internals.addCorsAndBearerToken
         }
       }
@@ -153,7 +42,7 @@ exports.register = function (plugin, options, next) {
       handler: {
         proxy: {
           passThrough: true,
-          mapUri: internals.mapProxyPath,
+          mapUri: mapProxyPath,
           onResponse: internals.addCorsAndBearerToken
         }
       }
@@ -164,7 +53,7 @@ exports.register = function (plugin, options, next) {
       handler: {
         proxy: {
           passThrough: true,
-          mapUri: internals.mapProxyPath,
+          mapUri: mapProxyPath,
           onResponse: internals.addCorsAndBearerToken
         }
       }
@@ -175,7 +64,7 @@ exports.register = function (plugin, options, next) {
       handler: {
         proxy: {
           passThrough: true,
-          mapUri: internals.mapProxyPath,
+          mapUri: mapProxyPath,
           onResponse: internals.addCorsAndBearerToken
         }
       }
@@ -186,7 +75,7 @@ exports.register = function (plugin, options, next) {
       handler: {
         proxy: {
           passThrough: true,
-          mapUri: internals.mapProxyPath,
+          mapUri: mapProxyPath,
           onResponse: internals.addCorsAndBearerToken
         }
       }
@@ -219,22 +108,22 @@ exports.register = function (plugin, options, next) {
     {
       method: 'GET',
       path: '/_api/_plugins/{name}/_api/{p*}',
-      handler: internals.handlePluginRequest
+      handler: handlePluginRequest
     },
     {
       method: 'PUT',
       path: '/_api/_plugins/{name}/_api/{p*}',
-      handler: internals.handlePluginRequest
+      handler: handlePluginRequest
     },
     {
       method: 'POST',
       path: '/_api/_plugins/{name}/_api/{p*}',
-      handler: internals.handlePluginRequest
+      handler: handlePluginRequest
     },
     {
       method: 'DELETE',
       path: '/_api/_plugins/{name}/_api/{p*}',
-      handler: internals.handlePluginRequest
+      handler: handlePluginRequest
     },
     {
       method: 'GET',
@@ -270,7 +159,7 @@ exports.register = function (plugin, options, next) {
     {
       method: 'GET',
       path: '/_api/_files/hoodie.js',
-      handler: internals.getHoodiePath
+      handler: getHoodiePath
     }
   ]);
 
@@ -280,6 +169,3 @@ exports.register = function (plugin, options, next) {
 exports.register.attributes = {
   pkg: require('./package.json')
 };
-
-//for unit tests:
-exports.internals = internals;
